Breakout visualization tests into multiple files
