+++ superset/assets/cypress/integration/explore/visualizations/line.js
@@ -1,43 +1,4 @@
-// ***********************************************
-// Tests for visualization types
-// ***********************************************
-
-const FORM_DATA_DEFAULTS = {
-  datasource: '3__table',
-  granularity_sqla: 'ds',
-  time_grain_sqla: null,
-  time_range: '100+years+ago+:+now',
-  adhoc_filters: [],
-  groupby: [],
-  limit: null,
-  timeseries_limit_metric: null,
-  order_desc: false,
-  contribution: false,
-};
-
-const NUM_METRIC = {
-    expressionType: 'SIMPLE',
-    column: {
-      id: 336,
-      column_name: 'num',
-      verbose_name: null,
-      description: null,
-      expression: '',
-      filterable: false,
-      groupby: false,
-      is_dttm: false,
-      type: 'BIGINT',
-      database_expression: null,
-      python_date_format: null,
-      optionName: '_col_num',
-    },
-    aggregate: 'SUM',
-    sqlExpression: null,
-    hasCustomLabel: false,
-    fromFormData: false,
-    label: 'Sum(num)',
-    optionName: 'metric_1de0s4viy5d_ly7y8k6ghvk',
-  };
+import { FORM_DATA_DEFAULTS, NUM_METRIC } from './shared.helper';
 
 describe('Line', function () {
   const LINE_CHART_DEFAULTS = { ...FORM_DATA_DEFAULTS, viz_type: 'line' };
@@ -50,7 +11,7 @@ describe('Line', function () {
 
     cy.route('POST', '/superset/explore_json/**').as('getJson');
     cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson' });
+    cy.verifySliceSuccess({ waitAlias: '@getJson', chartSelector: 'svg' });
   });
 
   it('Test line chart with groupby', function () {
@@ -64,7 +25,7 @@ describe('Line', function () {
 
     cy.route('POST', '/superset/explore_json/**').as('getJson');
     cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson' });
+    cy.verifySliceSuccess({ waitAlias: '@getJson', chartSelector: 'svg' });
   });
 
   it('Test line chart with simple filter', function () {
@@ -87,7 +48,7 @@ describe('Line', function () {
 
     cy.route('POST', '/superset/explore_json/**').as('getJson');
     cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson' });
+    cy.verifySliceSuccess({ waitAlias: '@getJson', chartSelector: 'svg' });
   });
 
   it('Test line chart with series limit sort asc', function () {
@@ -104,7 +65,7 @@ describe('Line', function () {
 
     cy.route('POST', '/superset/explore_json/**').as('getJson');
     cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson' });
+    cy.verifySliceSuccess({ waitAlias: '@getJson', chartSelector: 'svg' });
   });
 
   it('Test line chart with series limit sort desc', function () {
@@ -122,7 +83,7 @@ describe('Line', function () {
 
     cy.route('POST', '/superset/explore_json/**').as('getJson');
     cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson' });
+    cy.verifySliceSuccess({ waitAlias: '@getJson', chartSelector: 'svg' });
   });
 
   it('Test line chart with rolling avg', function () {
@@ -135,7 +96,7 @@ describe('Line', function () {
 
     cy.route('POST', '/superset/explore_json/**').as('getJson');
     cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson' });
+    cy.verifySliceSuccess({ waitAlias: '@getJson', chartSelector: 'svg' });
   });
 
   it('Test line chart with time shift 1 year', function () {
@@ -148,7 +109,7 @@ describe('Line', function () {
 
     cy.route('POST', '/superset/explore_json/**').as('getJson');
     cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson' });
+    cy.verifySliceSuccess({ waitAlias: '@getJson', chartSelector: 'svg' });
   });
 
   it('Test line chart with time shift yoy', function () {
@@ -161,7 +122,7 @@ describe('Line', function () {
 
     cy.route('POST', '/superset/explore_json/**').as('getJson');
     cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson' });
+    cy.verifySliceSuccess({ waitAlias: '@getJson', chartSelector: 'svg' });
   });
 
   it('Test line chart with time shift percentage change', function () {
@@ -174,62 +135,6 @@ describe('Line', function () {
 
     cy.route('POST', '/superset/explore_json/**').as('getJson');
     cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson' });
-  });
-});
-
-
-// Big Number Total
-
-describe('Big Number', function () {
-  const BIG_NUMBER_DEFAULTS = { ...FORM_DATA_DEFAULTS, viz_type: 'big_number_total' };
-
-  it('Test big number chart with adhoc metric', function () {
-    cy.server();
-    cy.login();
-
-    const formData = { ...BIG_NUMBER_DEFAULTS, metric: NUM_METRIC };
-
-    cy.route('POST', '/superset/explore_json/**').as('getJson');
-    cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson', querySubstring: NUM_METRIC.label, getSvg: false });
-  });
-
-  it('Test big number chart with simple filter', function () {
-    cy.server();
-    cy.login();
-
-    const filters = [{
-      expressionType: 'SIMPLE',
-      subject: 'name',
-      operator: 'in',
-      comparator: ['Aaron', 'Amy', 'Andrea'],
-      clause: 'WHERE',
-      sqlExpression: null,
-      fromFormData: true,
-      filterOptionName: 'filter_4y6teao56zs_ebjsvwy48c',
-    }];
-
-    const formData = { ...BIG_NUMBER_DEFAULTS, metric: 'count', adhoc_filters: filters };
-
-    cy.route('POST', '/superset/explore_json/**').as('getJson');
-    cy.visitChartByParams(JSON.stringify(formData));
-    cy.verifySliceSuccess({ waitAlias: '@getJson', getSvg: false });
-  });
-
-  it('Test big number chart ignores groupby', function () {
-    cy.server();
-    cy.login();
-
-    const formData = { ...BIG_NUMBER_DEFAULTS, metric: NUM_METRIC, groupby: ['state'] };
-
-    cy.route('POST', '/superset/explore_json/**').as('getJson');
-    cy.visitChartByParams(JSON.stringify(formData));
-    cy.wait(['@getJson']).then((data) => {
-      expect(data.status).to.eq(200);
-      expect(data.response.body).to.have.property('error', null);
-      expect(data.response.body.query).not.contains(formData.groupby[0]);
-      cy.get('.slice_container');
-    });
+    cy.verifySliceSuccess({ waitAlias: '@getJson', chartSelector: 'svg' });
   });
 });
